Added checkbox for images
